Applied skeletal implementation to interface migration automated refactoring.
